replace with long int

+++ mujoco_ros_control/src/mujoco_ros_control.cpp
@@ -107,7 +107,7 @@ void MujocoRosControl::init()
 void MujocoRosControl::update()
 {
   // get simulation time and period
-  int nanosec_time = (mujoco_data->time) * 1e9;
+  long int nanosec_time = (mujoco_data->time) * 1e9;
   ros::Time sim_time_ros(mujoco_data->time, nanosec_time); 
 
   ros::Duration sim_period = sim_time_ros - last_update_sim_time_ros_;
